refactor: signup api, utils.js
