+++ insta-clone/src/pages/Posts/CreatePostModal.jsx
@@ -1,3 +1,4 @@
+import apiClient from '@/services/apiClient';
 import { useState, useRef } from 'react';
 
 const CreatePostModal = ({ onClose }) => {
@@ -41,7 +42,7 @@ const CreatePostModal = ({ onClose }) => {
     if (!selectedImage) return;
 
     try {
-      const response = await fetch('/api/posts', {
+      const response = await apiClient.post('/api/posts', {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
@@ -74,32 +75,31 @@ const CreatePostModal = ({ onClose }) => {
       if (
         window.confirm('작성 중인 내용이 삭제됩니다. 그래도 나가시겠습니까?')
       ) {
-        onClose();  
-      } 
-    }else {
         onClose();
       }
-    
+    } else {
+      onClose();
+    }
   };
 
   return (
-    <div className="fixed inset-0 bg-black/70 flex items-center justify-center z-50"
-    onClick={handleCancel} >
-     <button
+    <div
+      className="fixed inset-0 bg-black/70 flex items-center justify-center z-50"
       onClick={handleCancel}
-      className="absolute top-3 right-3"
-      > 
-      <img 
-      src="/assets/icons/cancel.svg" 
-      className=" w-7 h-7 hover:opacity-60 transition-colors brightness-0 invert-[1]"  // invert 클래스 사용
-     />
+    >
+      <button onClick={handleCancel} className="absolute top-3 right-3">
+        <img
+          src="/assets/icons/cancel.svg"
+          className=" w-7 h-7 hover:opacity-60 transition-colors brightness-0 invert-[1]" // invert 클래스 사용
+        />
       </button>
-      
-      <div className="bg-white w-full max-w-[500px] rounded-xl h-[min(90vh,500px)] flex flex-col"
-      onClick={(e) => e.stopPropagation()}>
+
+      <div
+        className="bg-white w-full max-w-[500px] rounded-xl h-[min(90vh,500px)] flex flex-col"
+        onClick={e => e.stopPropagation()}
+      >
         {/* 헤더 */}
         <div className="relative flex items-center justify-center p-3 border-b border-gray-200">
-        
           <h2 className="text-base font-semibold">새 게시물 만들기</h2>
           {selectedImage && (
             <button
